requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
python-telegram-bot>=20.0
psutil>=5.9.0
pyyaml>=6.0
typing-extensions>=4.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
